refactor(router): reorganize metadata route types and imports
- Moved metadata route types from `../types/metadata-route` to a new `./metadata-route` file for better organization.
- Updated imports across the codebase to reflect the new file structure.
- Refactored server configuration types to be sourced from `./server-config` instead of `../types/server-config` for consistency.
