+++ packages/rari/src/index.ts
@@ -6,5 +6,6 @@ export { RariRequest } from './proxy/RariRequest'
 export { RariResponse } from './proxy/RariResponse'
 
 export type { ProxyConfig, ProxyFunction, RariFetchEvent, RariURL } from './proxy/types'
+
+export type { Feed, FeedEntry, Robots, RobotsRule, Sitemap, SitemapEntry, SitemapImage, SitemapVideo } from './router/metadata-route'
 export type { LayoutProps, Metadata, PageProps } from './router/types'
-export type { Feed, FeedEntry, Robots, RobotsRule, Sitemap, SitemapEntry, SitemapImage, SitemapVideo } from './types/metadata-route'
