+++ packages/rari/src/router/feed-generator.ts
@@ -1,4 +1,4 @@
-import type { Feed, FeedEntry } from '../types/metadata-route'
+import type { Feed, FeedEntry } from './metadata-route'
 import { Buffer } from 'node:buffer'
 import { promises as fs } from 'node:fs'
 import path from 'node:path'
