+++ packages/rari/src/router/metadata-route.ts
