+++ packages/rari/src/router/robots-generator.ts
@@ -1,4 +1,4 @@
-import type { Robots } from '../types/metadata-route'
+import type { Robots } from './metadata-route'
 import { Buffer } from 'node:buffer'
 import { promises as fs } from 'node:fs'
 import path from 'node:path'
