+++ packages/rari/src/vite.ts
@@ -32,6 +32,8 @@ export { rariProxy } from './proxy/vite-plugin'
 
 export type { ProxyPluginOptions } from './proxy/vite-plugin'
 
+export type { Robots, RobotsRule, Sitemap, SitemapEntry, SitemapImage, SitemapVideo } from './router/metadata-route'
+
 export {
   clearPropsCache,
   clearPropsCacheForComponent,
@@ -74,7 +76,9 @@ export type { Metadata } from './router/types'
 
 export { rariRouter } from './router/vite-plugin'
 
-export type { Robots, RobotsRule, Sitemap, SitemapEntry, SitemapImage, SitemapVideo } from './types/metadata-route'
+export { defineRariConfig, defineRariOptions } from './vite/index'
+
+export type { RariOptions, RouterPluginOptions } from './vite/index'
 
 export type {
   ServerCacheConfig,
@@ -83,11 +87,7 @@ export type {
   ServerConfig,
   ServerCSPConfig,
   ServerUseCacheConfig,
-} from './types/server-config'
-
-export { defineRariConfig, defineRariOptions } from './vite/index'
-
-export type { RariOptions, RouterPluginOptions } from './vite/index'
+} from './vite/server-config'
 
 export function rari(options?: Parameters<typeof _rari>[0]): any[] {
   return _rari(options)
