+++ test/unit/router/robots-generator.test.ts
@@ -1,4 +1,4 @@
-import type { Robots } from '@rari/types/metadata-route'
+import type { Robots } from '@rari/router/metadata-route'
 import { generateRobotsTxt } from '@rari/router/robots-generator'
 import { describe, expect, it } from 'vite-plus/test'
